backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.2.1
pandas==2.2.0
openpyxl==3.1.2
pyyaml==6.0.1
faker==23.2.1
xmltodict==0.13.0
pyarrow==15.0.0
httpx==0.27.0
anthropic==0.49.0
openai==1.40.0
google-generativeai==0.5.0
greenlet==3.0.3
